move the women watches routes to diffrent page to make index.js to crouded
