+++ app.js
@@ -15,7 +15,8 @@ var MongoStore = require('connect-mongo')(session);
 var routes = require('./routes/index');
 var users = require('./routes/users');
 var admin = require('./routes/admin');
-var men = require('./routes/men')
+var men = require('./routes/men');
+var women = require('./routes/women');
 
 var app = express();
 
@@ -55,7 +56,7 @@ app.use('/', routes);
 app.use('/users', users);
 app.use('/admin', admin);
 app.use('/Mens-Watches', men)
-
+app.use('/Womens-Watches',women)
 
 // catch 404 and forward to error handler
 app.use(function(req, res, next) {
