+++ routes/index.js
@@ -35,33 +35,7 @@ router.get('/Mens-watches-carts/:id', function(req, res, next){
     res.redirect('/')
   })
 });
-/////////////////////////////////////////////////////////////////////////////////
-/* GET mens watches page. */
-router.get('/Mens-Watches', function(req, res, next) {
-  Men.find(function(err, docs) {
-      var productChunks = [];
-      var chunkSize = 3;
-      for (var i = 0; i < docs.length; i += chunkSize) {
-          productChunks.push(docs.slice(i, i + chunkSize));
-      }
-      res.render('mens/index', { title: 'Cool Watches', products: productChunks });
-  })
-});
 
-/* GET mens watches show page. */
-router.get('/Mens-Watches/:id', function(req, res, next){
-  var product_id = req.param('id');
-  Men.findOne({'_id': product_id}, function(err, products){
-    if (!err){
-      var product = [];
-      var c = products;
-      product.push(c);
-      res.render('mens/show', {product: product});
-    }else{
-      return console.log(err);
-    }
-  });
-});
 
 /* GET mens watches  add to cart from the show page. */
 router.get('/mens-watches-cart/:id', function(req, res, next){
@@ -153,33 +127,8 @@ router.get('/kids-watches-cart/:id', function(req, res, next){
     res.redirect('/kids-Watches/'+ product.id)
   })
 });
-/////////////////////////////////////////////////////////////////////////////////
-/* GET women watches page. */
-router.get('/Womens-Watches', function(req, res, next) {
-  Women.find(function(err, docs) {
-      var productChunks = [];
-      var chunkSize = 3;
-      for (var i = 0; i < docs.length; i += chunkSize) {
-          productChunks.push(docs.slice(i, i + chunkSize));
-      }
-      res.render('women/index', { title: 'Cool Watches', products: productChunks });
-  })
-});
 
-/* GET mens watches show page. */
-router.get('/Womens-Watches/:id', function(req, res, next){
-  var product_id = req.param('id');
-  Women.findOne({'_id': product_id}, function(err, products){
-    if (!err){
-      var product = [];
-      var c = products;
-      product.push(c);
-      res.render('women/show', {product: product});
-    }else{
-      return console.log(err);
-    }
-  });
-});
+
 
 /* GET mens watches  add to cart from the show page. */
 router.get('/womens-watches-cart/:id', function(req, res, next){
